requirements.txt
requests>=2.31.0
pyyaml>=6.0
geopy>=2.4.0
